Adding a generic dropzone
